global-exception-handler

+++ src/main/java/com/me/odontologo/service/implementation/OdontologoService.java
@@ -1,6 +1,8 @@
 package com.me.odontologo.service.implementation;
 
 import com.me.odontologo.entity.Odontologo;
+import com.me.odontologo.exception.BadRequestException;
+import com.me.odontologo.exception.NoContentException;
 import com.me.odontologo.repository.IOdontologoRepository;
 import com.me.odontologo.service.IOdontologoService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,19 +20,48 @@ public class OdontologoService implements IOdontologoService {
 
     @Override
     public Odontologo guardarOdontologo(Odontologo odontologo){
-
-        return odontologoRepository.save(odontologo);
+        if(odontologo.getMatricula() != null){
+            return odontologoRepository.save(odontologo);
+        }else{
+            throw new BadRequestException("Falta el valor del campo matrícula");
+        }
     }
+
+    @Override
+    public Odontologo actualizarOdontologo(Odontologo odontologo){
+        if( odontologo.getId() != null
+                && odontologoRepository.findById(odontologo.getId()).isPresent()){
+            return odontologoRepository.save(odontologo);
+        }else{
+            throw new BadRequestException("El odontologo buscado para actualizar no existe");
+        }
+    }
+
     @Override
     public void eliminarOdontologo(Long id){
-        odontologoRepository.deleteById(id);
+        if(odontologoRepository.findById(id).isPresent()){
+            odontologoRepository.deleteById(id);
+        }else{
+            throw new BadRequestException("El odontologo no se puedo eliminar" +
+                    " porque no existe");
+        }
     }
     @Override
     public List<Odontologo> buscarTodosLosOdontologos(){
-        return odontologoRepository.findAll();
+        List<Odontologo> odontologos = odontologoRepository.findAll();
+        if (!odontologos.isEmpty()){
+            return odontologos;
+        }else{
+            throw new NoContentException();
+        }
     }
     @Override
     public Optional<Odontologo> buscar(Long id){
-        return odontologoRepository.findById(id);
+        Optional<Odontologo> odontologo = odontologoRepository.findById(id);
+        if(odontologo.isPresent()){
+            return odontologo;
+        }else{
+            throw new BadRequestException("El odontologo buscado no existe");
+        }
     }
 }
